Shuffle tests to their relevant new places
